ver 0.0.3.11 add FB meta, change way to get current user
