+++ schoolcms/static/schoolcms/component/component.jsx
